Use API Key ID in filename and delete API Key file in a clean up step
To avoid any conflicts, we now use the API Key ID to construct the name of the
API Key JSON file, and we save it inside of `Agent.BuildDirectory` or
`Agent.TempDirectory`.
By default, the API Key JSON file will be deleted after its use. That should
be much safer. Using the `DEBUG_API_KEY_FILE` environment variable keeps it
from being deleted. We use that env var for testing so we can verify that the
file has been created with the correct content.