Change `apiKeyFileName` to `apiKeyFilePath` and prefer Agent.TempDirectory
As per @egor-bryzgalov's code review, I've changed the name of the variable
holding the path to the API Key file to make it clearer that it's not just
the file *name*.

Also change where the file is saved on disk: Prefer Agent.TempDirectory
to Agent.BuildDirectory. This required some changes to the tests, since
the system also writes a `.taskkey` file into the temp dir, which we need
to clean up in order to delete the temp directory we had to create for
testing.